// src/eat-apples/logic/generated/org.accordproject.eatapples@0.2.0.ts
/* eslint-disable @typescript-eslint/no-empty-interface */
// Generated code for namespace: org.accordproject.eatapples@0.2.0

// imports
import {IClause} from './org.accordproject.contract@0.2.0';
import {IRequest,IResponse} from './org.accordproject.runtime@0.2.0';
import {IMonetaryAmount} from './org.accordproject.money@0.3.0';
import {IEvent} from './concerto@1.0.0';

// interfaces
export interface IFood extends IRequest {
   produce: string;
   price: IMonetaryAmount;
}

export interface IOutcome extends IResponse {
   notice: string;
}

export interface IBill extends IEvent {
   billTo: string;
   amount: IMonetaryAmount;
}

export interface ITemplateModel extends IClause {
   employee: string;
   company: string;
   tax: number;
}